requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
gunicorn==19.9.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
whitenoise==4.1.2
